+++ src/configs/keymap.rs
@@ -44,6 +44,7 @@ pub enum EditorAction {
     EndOfFile,
     StartOfLine,
     StartOfFile,
+    IdiomCommand,
     FindReferences,
     GoToDeclaration,
     Help,
@@ -100,6 +101,7 @@ pub struct EditorUserKeyMap {
     end_of_file: String,
     start_of_line: String,
     start_of_file: String,
+    idiom_command: String,
     find_references: String,
     go_to_declaration: String,
     help: String,
@@ -156,6 +158,7 @@ impl From<EditorUserKeyMap> for HashMap<KeyEvent, EditorAction> {
         insert_key_event(&mut hash, &val.end_of_file, EditorAction::EndOfFile);
         insert_key_event(&mut hash, &val.start_of_line, EditorAction::StartOfLine);
         insert_key_event(&mut hash, &val.start_of_file, EditorAction::StartOfFile);
+        insert_key_event(&mut hash, &val.idiom_command, EditorAction::IdiomCommand);
         insert_key_event(&mut hash, &val.find_references, EditorAction::FindReferences);
         insert_key_event(&mut hash, &val.go_to_declaration, EditorAction::GoToDeclaration);
         insert_key_event(&mut hash, &val.help, EditorAction::Help);
@@ -213,6 +216,7 @@ impl Default for EditorUserKeyMap {
             end_of_file: format!("{CTRL} && {END}"),
             start_of_line: HOME.to_owned(),
             start_of_file: format!("{CTRL} && {HOME}"),
+            idiom_command: format!("{CTRL} && ;"),
             find_references: format!("{F}9"),
             go_to_declaration: format!("{F}12"),
             help: format!("{F}1"),